+++ src/main/java/com/mygym/crm/backstages/repositories/services/TrainerService.java
@@ -3,9 +3,9 @@ package com.mygym.crm.backstages.repositories.services;
 
 import com.mygym.crm.backstages.domain.models.Trainer;
 
-public interface TrainerService<T> extends BaseService<Trainer, Integer> {
+public interface TrainerService<T> extends BaseService<Trainer, Long> {
     void create(T t);
 
-    void update(Integer id, T t);
+    void update(Long id, T t);
 
 }
